Fixed the bug and improved the api (and debugging) slightly.
I'm still not sure this is good approach. Off-by-one errors are really
easy to make. I think the issue is we want to talk about the location of
a placeholder, but our only tool is the size a function was, and those
don't place nicely with half-open intervals that lead to simple-ish
arithmetic.