Reimplement ValueStack.
By effecitvely implementing our own vector and having an understanding
of which instructions add or remove from the stack, we can avoid
reallocation in cases where we know we're not adding to the stack. This
means that in those circumstances we don't have to save any stack state
(on most instructions) and this saves ~3% in the fibonacci benchmark.