Save another ~4% on fib(38) benchmark.
On Apple M1 Pro, recursive fib(38) now takes ~1.2497s.

The bulk of this seems to be from inlining std::memcpy with the default
implementation. Even though the optimizer sees through this and
replaces it with a call to std::memcpy, it manages to avoid saving and
restoring two registers in the process, which I suspect is worth some of
the savings.

A more negligible portion of the savings is due to the change from
ValueStack to passing ValueStackRef. This is maybe ~1% of the 4%
savings. There are still many registers being saved on these codepaths
that I believe don't need to be. Looking at the assembly for
`jasmin::Push` vs `jasmin::Swap`, all of these register manipulations
seem to be elided from the latter but not the former. This likely
indicates something about the reallocation function calls blocking the
compiler from eliding the register manipulation, since that's
fundamentally the difference between the instructions.