Saving another 8% on fib(38) benchmark.
Recursive fib(38) currently takes 1.146s on Apple M1 Pro.

This works by having Push be builtin so that we can make reallocation
part happen behind a tail-call. I believe the speedup comes from
allowing us to avoid saving registers in the common case where we don't
need to reallocate. The registers do need to be saved before a
reallocation, buth that can be scoped to a smaller function that only
occurs in the unlikely case of reallocation.