Save another 3.5% on recursive fib(38) benchmark.
Now 1.0304s on Apple M1 Pro. The technique here is similarly scoping the
register savings down to only when a reallocation actually occurs for
the call stack. This a better optimization on function-call-heavy code
(e.g.recursive functions like this benchmark), so for most functions,
this is probably in the noise.